styled question_metadata I changed the structure for the likes and disklikes icons (the metadata). There I changed from just <i> tag to list <li> and <a> (I need to delete the <a> cause these are links).
I deleted the list-style from html with li{
list-style: none;>

And styled the question_content with a padding and display flex so the metadata is on the left side and the header is on the right side.
After that I changed the .question_metadata to grid and the template-columns to 2(so two columns in a row) and styled the icon. The last code in css is question_meta is that the icon is above the numbers and I also changed the size and color.