added question_date and styled it
On top I changed the .question_title to .main_title cuase that's a better name.
Due that we have a new div with the date I decided to use flexbox and set the justify-content to space-between so the tags are on the eft side and the date is on the right side.