created root.css where every variable is stored I created a root css file where every variable is stored in so I do not have to set the variables for every file new.
I thought about having a css file for every site. So having a seperate css file for /home, for /login, for /signup and so on. But the problem is that we need to declare the variables for every file and I also wanted to have the nav-bar same for every site. So I created a root.css where the nav-bar and all of the variables are stored (root-color).
So now I just need to set a new color for the variable root-color and every color will change, convinient.
Do not hang up with a problem, solve it.