created an automatic textarea resizer with js Added nice code lines from a nice guy on youtube https://youtu.be/Yor9Y73M764.
First we add a new constant from html called textarea. We are getting all textaread in the whole html, luckily we have just one textarea so thta is not problem, even if we had more thane one textarea the resize function is great. After that we get the estimated height with the arrow function.
If we console.log the scheight out it will give us the height information. In the last line we resize it with scHeight in px.